Add ErrorDomainMapper.kt Refactor tests
