+++ app/src/main/java/com/jeanbernad/randomuser/core/DependencyProvider.kt
@@ -7,10 +7,12 @@ import com.google.gson.GsonBuilder
 import com.jeanbernad.randomuser.data.user.*
 import com.jeanbernad.randomuser.data.user.local.*
 import com.jeanbernad.randomuser.data.user.remote.UserService
+import com.jeanbernad.randomuser.domain.ErrorDomainMapper
 import com.jeanbernad.randomuser.domain.user.BaseUserDataToDomainMapper
 import com.jeanbernad.randomuser.domain.user.UserDomain
 import com.jeanbernad.randomuser.domain.user.UserDomainToPresentationMapper
 import com.jeanbernad.randomuser.domain.user.UserInteractor
+import com.jeanbernad.randomuser.presentation.ErrorPresentationMapper
 import com.jeanbernad.randomuser.presentation.user.BaseUserDomainToPresentationMapper
 import com.jeanbernad.randomuser.presentation.user.UserPresentationModel
 import com.jeanbernad.randomuser.presentation.user.UserViewModel
