+++ app/src/main/java/com/jeanbernad/randomuser/presentation/user/BaseUserDomainToPresentationMapper.kt
@@ -1,6 +1,6 @@
 package com.jeanbernad.randomuser.presentation.user
 
-import com.jeanbernad.randomuser.core.ErrorPresentationMapper
+import com.jeanbernad.randomuser.presentation.ErrorPresentationMapper
 import com.jeanbernad.randomuser.core.ErrorType
 import com.jeanbernad.randomuser.domain.user.UserDomainToPresentationMapper
