+++ app/src/test/java/com/jeanbernad/randomuser/data/user/UserRepositoryTest.kt
@@ -1,6 +1,6 @@
 package com.jeanbernad.randomuser.data.user
 
-import com.jeanbernad.randomuser.core.ErrorDomainMapper
+import com.jeanbernad.randomuser.domain.ErrorDomainMapper
 import com.jeanbernad.randomuser.data.user.remote.UserRemoteModel
 import com.jeanbernad.randomuser.data.user.remote.entity.result.*
 import com.jeanbernad.randomuser.data.user.remote.entity.result.location.Coordinates
